requirements.txt
discord.py
requests
jishaku
pip
aiohttp
uptime
